Refactored and added camera code.
TONS of stuff changed in this commit:

Refactored the package names to make more sense.  Now everything is under
Alkehestry Engine and it has sub packages of graphics, game_engine, etc.

Added code to handle scrolling of the map and and general movement.
Movement itself needs a bit more work.  Boundaries should be working with
this commit.